refactor: common 공용 폴더 적용
