+++ src/domain/auth/page/AdminLoginCallbackPage.tsx
@@ -1,8 +1,8 @@
 import axios from "axios";
 import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import { PATH } from "../../../constant/Path";
-import { API_PATH } from "../../../constant/ApiPath";
+import { PATH } from "../../../common/constant/Path";
+import { API_PATH } from "../../../common/constant/ApiPath";
 import { StoredMemberInfo } from "../type/AuthType";
 
 export default function AdminLoginCallbackPage() {
