+++ src/domain/folder/component/setting-folder/FolderBox.tsx
@@ -1,5 +1,5 @@
-import { useWebHookModalStore } from "../../../../store/webHookModalStore";
-import { Icon } from "../../../common/component/Icon";
+import { useWebHookModalStore } from "../../../../common/store/webHookModalStore";
+import { Icon } from "../../../../common/component/Icon";
 import { Folder } from "../../../layout/type/SideBarType";
 
 type Props = {
